+++ src/main/java/com/example/UserLoginService/Events/UserCreatedEvent.java
@@ -5,18 +5,18 @@ import java.time.LocalDate;
 
 public class UserCreatedEvent implements Serializable
 {
-    private int Id;
+    private Long Id;
 
     private String PassWord;
 
     private String Email;
 
-    private int RoleId;
+    private Long RoleId;
     private String RoleName;
 
     private LocalDate CreatedAt;
 
-    public UserCreatedEvent(int id, String passWord, String email, int roleId, String roleName, LocalDate createdAt) {
+    public UserCreatedEvent(Long id, String passWord, String email, Long roleId, String roleName, LocalDate createdAt) {
         Id = id;
         PassWord = passWord;
         Email = email;
@@ -29,7 +29,7 @@ public class UserCreatedEvent implements Serializable
 
     }
 
-    public int getId() {
+    public Long getId() {
         return Id;
     }
 
@@ -41,7 +41,7 @@ public class UserCreatedEvent implements Serializable
         return Email;
     }
 
-    public int getRoleId() {
+    public Long getRoleId() {
         return RoleId;
     }
